requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.2
geographiclib==2.0
geopy==2.4.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
JPype1==1.5.0
MarkupSafe==2.1.5
packaging==23.2
pycryptodome==3.20.0
requests==2.31.0
urllib3==2.2.1
Werkzeug==3.0.1
